crypto-asymmetric/requirements.txt
pwntools=4.14.1
pycryptodome=3.23.0
factordb-python=1.3.0
gmpy2=2.2.1